+++ adapters.h
@@ -1,0 +1,65 @@
+#ifndef ADAPTERS_H
+#define ADAPTERS_H
+
+#ifndef _DEFAULT_SOURCE
+#define _DEFAULT_SOURCE
+#endif
+
+#include <iostream>
+#include <exception>
+#include <vector>
+#include <dlfcn.h>
+#include <fstream>
+#include <algorithm>
+#include <cstdio>
+#include <cerrno>
+#include <cstdlib>
+#include <cstring>
+#include <string>
+#include <memory>
+#include <cmath>
+#include <cstdarg>
+#include <stdint.h>
+#include <unistd.h>
+#include <dirent.h>
+#include <sys/types.h>
+#include <fcntl.h>
+#include <CL/cl.h>
+
+extern "C" {
+#include <pci/pci.h>
+}
+
+#ifdef __linux__
+#define LINUX 1
+#endif
+#include "./dependencies/ADL_SDK_V10.2/include/adl_sdk.h"
+
+#include "error.h"
+#include "atiadlhandle.h"
+#include "adlmaincontrol.h"
+#include "amdgpuadapterinfo.h"
+
+class Adapters
+{
+private:
+
+public:
+
+    static void ParseAdaptersList(const char* string, std::vector<int>& adapters, bool& allAdapters);
+
+    static void PrintAdaptersInfoVerbose(ADLMainControl& mainControl, int adaptersNum, const std::vector<int>& activeAdapters,
+                                         const std::vector<int>& choosenAdapters, bool useChoosen);
+
+    static void PrintAdaptersInfo(ADLMainControl& mainControl, int adaptersNum, const std::vector<int>& activeAdapters, const std::vector<int>& choosenAdapters,
+                                  bool useChoosen);
+
+    static void GetActiveAdaptersIndices(ADLMainControl& mainControl, int adaptersNum, std::vector<int>& activeAdapters);
+
+    static void PrintAdaptersInfoVerbose(AMDGPUAdapterHandle& handle, const std::vector<int>& choosenAdapters, bool useChoosen);
+
+    static void printAdaptersInfo(AMDGPUAdapterHandle& handle, const std::vector<int>& choosenAdapters, bool useChoosen);
+
+};
+
+#endif /* ADAPTERS_H */
