+++ amdcovc.cpp
@@ -47,6 +47,7 @@ extern "C" {
 #ifdef __linux__
 #define LINUX 1
 #endif
+
 #include "./dependencies/ADL_SDK_V10.2/include/adl_sdk.h"
 
 #include "error.h"
@@ -55,1330 +56,14 @@ extern "C" {
 #include "amdgpuadapterhandle.h"
 #include "amdgpuadapterinfo.h"
 #include "pciaccess.h"
+#include "adapters.h"
+#include "parameters.h"
 
 #define AMDCOVC_VERSION "0.4.0"
 
 pci_access* pciAccess = nullptr;
 
 pci_filter pciFilter;
-
-static void printAdaptersInfo(AMDGPUAdapterHandle& handle, const std::vector<int>& choosenAdapters, bool useChoosen)
-{
-    int adaptersNum = handle.getAdaptersNum();
-    auto choosenIter = choosenAdapters.begin();
-    int i = 0;
-
-    for (int ai = 0; ai < adaptersNum; ai++)
-    {
-        if (useChoosen && (choosenIter==choosenAdapters.end() || *choosenIter != i ))
-        {
-            i++;
-            continue;
-        }
-
-        const AMDGPUAdapterInfo adapterInfo = handle.parseAdapterInfo(ai);
-
-        std::cout << "Adapter " << i << ": " << adapterInfo.name << "\n  Core: " << adapterInfo.coreClock << " MHz, Mem: " <<
-                adapterInfo.memoryClock << " MHz, CoreOD: " << adapterInfo.coreOD << ", MemOD: " << adapterInfo.memoryOD << ", ";
-
-        if (adapterInfo.gpuLoad>=0)
-        {
-            std::cout << "Load: " << adapterInfo.gpuLoad << "%, ";
-        }
-
-        std::cout << "Temp: " << adapterInfo.temperature/1000.0 << " C, Fan: " <<
-                double(adapterInfo.fanSpeed-adapterInfo.minFanSpeed) / double(adapterInfo.maxFanSpeed - adapterInfo.minFanSpeed) * 100.0 <<
-                "%" << std::endl;
-
-        if (!adapterInfo.coreClocks.empty())
-        {
-            std::cout << "  Core clocks:";
-
-            for (uint32_t v: adapterInfo.coreClocks)
-            {
-                std::cout << " " << v;
-            }
-
-            std::cout << std::endl;
-        }
-
-        if (!adapterInfo.memoryClocks.empty())
-        {
-            std::cout << "  Memory Clocks: ";
-
-            for (uint32_t v: adapterInfo.memoryClocks)
-            {
-                std::cout << " " << v;
-            }
-
-            std::cout << std::endl;
-        }
-
-        if (useChoosen)
-        {
-            ++choosenIter;
-        }
-
-        i++;
-    }
-}
-
-static void printAdaptersInfoVerbose(AMDGPUAdapterHandle& handle, const std::vector<int>& choosenAdapters, bool useChoosen)
-{
-    int adaptersNum = handle.getAdaptersNum();
-    auto choosenIter = choosenAdapters.begin();
-    int i = 0;
-
-    for (int ai = 0; ai < adaptersNum; ai++)
-    {
-        if (useChoosen && (choosenIter == choosenAdapters.end() || *choosenIter != i ) )
-        {
-            i++;
-            continue;
-        }
-
-        const AMDGPUAdapterInfo adapterInfo = handle.parseAdapterInfo(ai);
-
-        std::cout << "Adapter " << i << ": " << adapterInfo.name << "\n"
-            "  Device Topology: " << adapterInfo.busNo << ':' << adapterInfo.deviceNo << ":" << adapterInfo.funcNo << "\n"
-            "  Vendor ID: " << adapterInfo.vendorId << "\n"
-            "  Device ID: " << adapterInfo.deviceId << "\n"
-            "  Current CoreClock: " << adapterInfo.coreClock << " MHz\n"
-            "  Current MemoryClock: " << adapterInfo.memoryClock << " MHz\n"
-            "  Core Overdrive: " << adapterInfo.coreOD << "\n"
-            "  Memory Overdrive: " << adapterInfo.memoryOD << "\n";
-
-        if (adapterInfo.gpuLoad>=0)
-        {
-            std::cout << "  GPU Load: " << adapterInfo.gpuLoad << "%\n";
-        }
-
-        std::cout << "  Current BusSpeed: " << adapterInfo.busSpeed << "\n"
-            "  Current BusLanes: " << adapterInfo.busLanes << "\n"
-            "  Temperature: " << adapterInfo.temperature/1000.0 << " C\n"
-            "  Critical temperature: " << adapterInfo.tempCritical/1000.0 << " C\n"
-            "  FanSpeed Min (Value): " << adapterInfo.minFanSpeed << "\n"
-            "  FanSpeed Max (Value): " << adapterInfo.maxFanSpeed << "\n"
-            "  Current FanSpeed: " <<
-                (double(adapterInfo.fanSpeed-adapterInfo.minFanSpeed)/ double(adapterInfo.maxFanSpeed-adapterInfo.minFanSpeed)*100.0) << "%\n"
-            "  Controlled FanSpeed: " << ( adapterInfo.defaultFanSpeed ? "yes" : "no" ) << "\n";
-
-        // print available core clocks
-        if (!adapterInfo.coreClocks.empty())
-        {
-            std::cout << "  Core clocks:\n";
-
-            for (uint32_t v: adapterInfo.coreClocks)
-            {
-                std::cout << "    " << v << "MHz\n";
-            }
-        }
-
-        if (!adapterInfo.memoryClocks.empty())
-        {
-            std::cout << "  Memory Clocks:\n";
-
-            for (uint32_t v: adapterInfo.memoryClocks)
-            {
-                std::cout << "    " << v << "MHz\n";
-            }
-        }
-
-        if (useChoosen)
-        {
-            ++choosenIter;
-        }
-
-        i++;
-        std::cout << std::endl;
-    }
-}
-
-/* AMDGPU code */
-
-static void getActiveAdaptersIndices(ADLMainControl& mainControl, int adaptersNum, std::vector<int>& activeAdapters)
-{
-    activeAdapters.clear();
-
-    for (int i = 0; i < adaptersNum; i++)
-    {
-        if (mainControl.isAdapterActive(i))
-        {
-            activeAdapters.push_back(i);
-        }
-    }
-}
-
-static void printAdaptersInfo(ADLMainControl& mainControl, int adaptersNum, const std::vector<int>& activeAdapters, const std::vector<int>& choosenAdapters,
-                              bool useChoosen)
-{
-    std::unique_ptr<AdapterInfo[]> adapterInfos(new AdapterInfo[adaptersNum]);
-    ::memset(adapterInfos.get(), 0, sizeof(AdapterInfo)*adaptersNum);
-    mainControl.getAdapterInfo(adapterInfos.get());
-
-    int i = 0;
-    auto choosenIter = choosenAdapters.begin();
-
-    for (int ai = 0; ai < adaptersNum; ai++)
-    {
-        if (!mainControl.isAdapterActive(ai))
-        {
-            continue;
-        }
-
-        if (useChoosen && (choosenIter==choosenAdapters.end() || *choosenIter!=i))
-        {
-            i++;
-            continue;
-        }
-
-        if (adapterInfos[ai].strAdapterName[0] == 0)
-        {
-            PCIAccess::GetFromPCI(adapterInfos[ai].iAdapterIndex, adapterInfos[ai]);
-        }
-
-        ADLPMActivity activity;
-        mainControl.getCurrentActivity(ai, activity);
-
-        std::cout << "Adapter " << i << ": " << adapterInfos[ai].strAdapterName << "\n"
-                "  Core: " << activity.iEngineClock/100.0 << " MHz, "
-                "Mem: " << activity.iMemoryClock/100.0 << " MHz, "
-                "Vddc: " << activity.iVddc/1000.0 << " V, "
-                "Load: " << activity.iActivityPercent << "%, "
-                "Temp: " << mainControl.getTemperature(ai, 0)/1000.0 << " C, "
-                "Fan: " << mainControl.getFanSpeed(ai, 0) << "%" << std::endl;
-
-        ADLODParameters odParams;
-        mainControl.getODParameters(ai, odParams);
-
-        std::cout << "  Max Ranges: Core: " << odParams.sEngineClock.iMin/100.0 << " - " << odParams.sEngineClock.iMax/100.0 << " MHz, "
-            "Mem: " << odParams.sMemoryClock.iMin/100.0 << " - " << odParams.sMemoryClock.iMax/100.0 << " MHz, " <<
-            "Vddc: " <<  odParams.sVddc.iMin/1000.0 << " - " << odParams.sVddc.iMax/1000.0 << " V\n";
-
-        int levelsNum = odParams.iNumberOfPerformanceLevels;
-
-        std::unique_ptr<ADLODPerformanceLevel[]> odPLevels(new ADLODPerformanceLevel[levelsNum]);
-
-        mainControl.getODPerformanceLevels(ai, false, levelsNum, odPLevels.get());
-
-        std::cout << "  PerfLevels: Core: " << odPLevels[0].iEngineClock/100.0 << " - " << odPLevels[levelsNum-1].iEngineClock/100.0 << " MHz, "
-            "Mem: " << odPLevels[0].iMemoryClock/100.0 << " - " << odPLevels[levelsNum-1].iMemoryClock/100.0 << " MHz, "
-            "Vddc: " << odPLevels[0].iVddc/1000.0 << " - " << odPLevels[levelsNum-1].iVddc/1000.0 << " V\n";
-
-        if (useChoosen)
-        {
-            ++choosenIter;
-        }
-
-        i++;
-        std::cout << std::endl;
-    }
-}
-
-static void printAdaptersInfoVerbose(ADLMainControl& mainControl, int adaptersNum, const std::vector<int>& activeAdapters, const std::vector<int>& choosenAdapters,
-                                     bool useChoosen)
-{
-    std::unique_ptr<AdapterInfo[]> adapterInfos(new AdapterInfo[adaptersNum]);
-    ::memset(adapterInfos.get(), 0, sizeof(AdapterInfo)*adaptersNum);
-
-    mainControl.getAdapterInfo(adapterInfos.get());
-
-    int i = 0;
-    auto choosenIter = choosenAdapters.begin();
-
-    for (int ai = 0; ai < adaptersNum; ai++)
-    {
-        if (!mainControl.isAdapterActive(ai))
-        {
-            continue;
-        }
-
-        if (useChoosen && (choosenIter==choosenAdapters.end() || *choosenIter!=i))
-        {
-            i++;
-            continue;
-        }
-
-        if (adapterInfos[ai].strAdapterName[0] == 0)
-        {
-            PCIAccess::GetFromPCI(adapterInfos[ai].iAdapterIndex, adapterInfos[ai]);
-        }
-
-        std::cout <<
-            "Adapter " << i << ": " << adapterInfos[ai].strAdapterName << "\n"
-            "  Device Topology: " << adapterInfos[ai].iBusNumber << ':' << adapterInfos[ai].iDeviceNumber << ":" << adapterInfos[ai].iFunctionNumber << "\n"
-            "  Vendor ID: " << adapterInfos[ai].iVendorID << std::endl;
-
-        ADLFanSpeedInfo fsInfo;
-        ADLPMActivity activity;
-
-        mainControl.getCurrentActivity(ai, activity);
-
-        std::cout << "  Current CoreClock: " << activity.iEngineClock / 100.0 << " MHz\n"
-            "  Current MemoryClock: " << activity.iMemoryClock / 100.0 << " MHz\n"
-            "  Current Voltage: " << activity.iVddc / 1000.0 << " V\n"
-            "  GPU Load: " << activity.iActivityPercent << "%\n"
-            "  Current PerfLevel: " << activity.iCurrentPerformanceLevel << "\n"
-            "  Current BusSpeed: " << activity.iCurrentBusSpeed << "\n"
-            "  Current BusLanes: " << activity.iCurrentBusLanes<< "\n";
-
-        int temperature = mainControl.getTemperature(ai, 0);
-
-        std::cout << "  Temperature: " << temperature/1000.0 << " C\n";
-
-        mainControl.getFanSpeedInfo(ai, 0, fsInfo);
-
-        std::cout << "  FanSpeed Min: " << fsInfo.iMinPercent << "%\n"
-            "  FanSpeed Max: " << fsInfo.iMaxPercent << "%\n"
-            "  FanSpeed MinRPM: " << fsInfo.iMinRPM << " RPM\n"
-            "  FanSpeed MaxRPM: " << fsInfo.iMaxRPM << " RPM" << "\n";
-
-        std::cout << "  Current FanSpeed: " << mainControl.getFanSpeed(ai, 0) << "%\n";
-
-        ADLODParameters odParams;
-        mainControl.getODParameters(ai, odParams);
-
-        std::cout <<
-            "  CoreClock: " << odParams.sEngineClock.iMin / 100.0 << " - " << odParams.sEngineClock.iMax / 100.0 <<
-            " MHz, step: " << odParams.sEngineClock.iStep / 100.0 << " MHz\n"
-            "  MemClock: " << odParams.sMemoryClock.iMin / 100.0 << " - " << odParams.sMemoryClock.iMax / 100.0 <<
-            " MHz, step: " << odParams.sMemoryClock.iStep / 100.0 << " MHz\n"
-            "  Voltage: " << odParams.sVddc.iMin / 1000.0 << " - " << odParams.sVddc.iMax / 1000.0 <<
-            " V, step: " << odParams.sVddc.iStep / 1000.0 << " V\n";
-
-        std::unique_ptr<ADLODPerformanceLevel[]> odPLevels(new ADLODPerformanceLevel[odParams.iNumberOfPerformanceLevels]);
-
-        mainControl.getODPerformanceLevels(ai, false, odParams.iNumberOfPerformanceLevels, odPLevels.get());
-
-        std::cout << "  Performance levels: " << odParams.iNumberOfPerformanceLevels << "\n";
-
-        for (int j = 0; j < odParams.iNumberOfPerformanceLevels; j++)
-        {
-            std::cout <<
-                "    Performance Level: " << j << "\n"
-                "      CoreClock: " << odPLevels[j].iEngineClock / 100.0 << " MHz\n"
-                "      MemClock: " << odPLevels[j].iMemoryClock / 100.0 << " MHz\n"
-                "      Voltage: " << odPLevels[j].iVddc / 1000.0 << " V\n";
-        }
-
-        mainControl.getODPerformanceLevels(ai, true, odParams.iNumberOfPerformanceLevels, odPLevels.get());
-
-        std::cout << "  Default Performance levels: " << odParams.iNumberOfPerformanceLevels << "\n";
-
-        for (int j = 0; j < odParams.iNumberOfPerformanceLevels; j++)
-        {
-            std::cout <<
-                "    Performance Level: " << j << "\n"
-                "      CoreClock: " << odPLevels[j].iEngineClock / 100.0 << " MHz\n"
-                "      MemClock: " << odPLevels[j].iMemoryClock / 100.0 << " MHz\n"
-                "      Voltage: " << odPLevels[j].iVddc / 1000.0 << " V\n";
-        }
-
-        std::cout.flush();
-
-        if (useChoosen)
-        {
-            ++choosenIter;
-        }
-
-        i++;
-        std::cout << std::endl;
-    }
-}
-
-static void parseAdaptersList(const char* string, std::vector<int>& adapters, bool& allAdapters)
-{
-    adapters.clear();
-    allAdapters = false;
-
-    if (::strcmp(string, "all") == 0)
-    {
-        allAdapters = true;
-        return;
-    }
-
-    while (true)
-    {
-        char* endptr;
-        errno = 0;
-        int adapterIndex = strtol(string, &endptr, 10);
-
-        if (errno!=0 || endptr==string)
-        {
-            throw Error("Unable to parse adapter index");
-        }
-
-        string = endptr;
-
-        if (*string == '-')
-        {
-            // if range
-            string++;
-            errno = 0;
-            int adapterIndexEnd = strtol(string, &endptr, 10);
-
-            if (errno!=0 || endptr==string)
-            {
-                throw Error("Unable to parse adapter index");
-            }
-
-            string = endptr;
-
-            if (adapterIndex>adapterIndexEnd)
-            {
-                throw Error("Wrong range of adapter indices in adapter list");
-            }
-
-            for (int i = adapterIndex; i <= adapterIndexEnd; i++)
-            {
-                adapters.push_back(i);
-            }
-        }
-        else
-        {
-            adapters.push_back(adapterIndex);
-        }
-
-        if (*string == 0)
-        {
-            break;
-        }
-
-        if (*string==',')
-        {
-            string++;
-        }
-        else
-        {
-            throw Error("Invalid data in adapter list");
-        }
-    }
-
-    std::sort(adapters.begin(), adapters.end());
-
-    adapters.resize(std::unique(adapters.begin(), adapters.end()) - adapters.begin());
-}
-
-enum class OVCParamType
-{
-    CORE_CLOCK,
-    MEMORY_CLOCK,
-    VDDC_VOLTAGE,
-    FAN_SPEED,
-    CORE_OD,
-    MEMORY_OD
-};
-
-enum: int
-{
-    LAST_PERFLEVEL = -1
-};
-
-struct OVCParameter
-{
-    OVCParamType type;
-    std::vector<int> adapters;
-    bool allAdapters;
-    int partId;
-    double value;
-    bool useDefault;
-    std::string argText;
-};
-
-static bool parseOVCParameter(const char* string, OVCParameter& param)
-{
-    const char* afterName = strchr(string, ':');
-
-    if (afterName==nullptr)
-    {
-        afterName = strchr(string, '=');
-
-        if (afterName==nullptr)
-        {
-            std::cerr << "This is not parameter: '" << string << "'!" << std::endl;
-
-            return false;
-        }
-    }
-
-    std::string name(string, afterName);
-    param.argText = string;
-    param.adapters.clear();
-    param.adapters.push_back(0); // default is 0
-    param.allAdapters = false;
-    param.partId = 0;
-    param.useDefault = false;
-    bool partIdSet = false;
-
-    if (name=="coreclk")
-    {
-        param.type = OVCParamType::CORE_CLOCK;
-        param.partId = LAST_PERFLEVEL;
-    }
-    else if (name=="memclk")
-    {
-        param.type = OVCParamType::MEMORY_CLOCK;
-        param.partId = LAST_PERFLEVEL;
-    }
-    else if (name=="coreod")
-    {
-        param.type = OVCParamType::CORE_OD;
-        param.partId = LAST_PERFLEVEL;
-    }
-    else if (name=="memod")
-    {
-        param.type = OVCParamType::MEMORY_OD;
-        param.partId = LAST_PERFLEVEL;
-    }
-    else if (name=="vcore")
-    {
-        param.type = OVCParamType::VDDC_VOLTAGE;
-        param.partId = LAST_PERFLEVEL;
-    }
-    else if (name=="fanspeed")
-    {
-        param.type = OVCParamType::FAN_SPEED;
-        partIdSet = false;
-    }
-    else if (name=="icoreclk")
-    {
-        param.type = OVCParamType::CORE_CLOCK;
-        partIdSet = true;
-    }
-    else if (name=="imemclk")
-    {
-        param.type = OVCParamType::MEMORY_CLOCK;
-        partIdSet = true;
-    }
-    else if (name=="ivcore")
-    {
-        param.type = OVCParamType::VDDC_VOLTAGE;
-        partIdSet = true;
-    }
-    else
-    {
-        std::cout << "Wrong parameter name in '" << string << "'!" << std::endl;
-        return false;
-    }
-
-    char* next;
-
-    if (*afterName == ':')
-    {
-        // if is
-        afterName++;
-
-        try
-        {
-            const char* afterList = ::strchr(afterName, ':');
-
-            if (afterList==nullptr)
-            {
-                afterList = ::strchr(afterName, '=');
-            }
-
-            if (afterList==nullptr)
-            {
-                afterList += strlen(afterName); // to end
-            }
-
-            if (afterList!=afterName)
-            {
-                std::string listString(afterName, afterList);
-                parseAdaptersList(listString.c_str(), param.adapters, param.allAdapters);
-                afterName = afterList;
-            }
-        }
-        catch(const Error& error)
-        {
-            std::cerr << "Unable to parse adapter list for '" << string << "': " << error.what() << std::endl;
-            return false;
-        }
-    }
-    else if (*afterName==0)
-    {
-        std::cerr << "Unterminated parameter '" << string << "'!" << std::endl;
-        return false;
-    }
-
-    if (*afterName==':' && !partIdSet)
-    {
-        afterName++;
-        errno = 0;
-        int value = strtol(afterName, &next, 10);
-
-        if (errno!=0)
-        {
-            std::cerr << "Unable to parse partId in '" << string << "'!" << std::endl;
-            return false;
-        }
-
-        if (afterName != next)
-        {
-            param.partId = value;
-        }
-
-        afterName = next;
-    }
-    else if (*afterName==0)
-    {
-        std::cerr << "Unterminated parameter '" << string << "'!" << std::endl;
-        return false;
-    }
-
-    if (*afterName=='=')
-    {
-        afterName++;
-        errno = 0;
-
-        if (::strcmp(afterName, "default") == 0)
-        {
-            param.useDefault = true;
-            afterName += 7;
-        }
-        else
-        {
-            param.value = strtod(afterName, &next);
-
-            if (errno!=0 || afterName==next)
-            {
-                std::cerr << "Unable to parse value in '" << string << "'!" << std::endl;
-                return false;
-            }
-            if (std::isinf(param.value) || std::isnan(param.value))
-            {
-                std::cerr << "Value of '" << string << "' is not finite!" << std::endl;
-                return false;
-            }
-            afterName = next;
-        }
-        if (*afterName!=0)
-        {
-            std::cerr << "Garbage in '" << string << "'!" << std::endl;
-            return false;
-        }
-    }
-    else
-    {
-        std::cerr << "Unterminated parameter '" << string << "'!" << std::endl;
-        return false;
-    }
-
-    return true;
-}
-
-struct FanSpeedSetup
-{
-    double value;
-    bool useDefault;
-    bool isSet;
-};
-
-struct AdapterIterator
-{
-    const std::vector<int>& adapters;
-    bool allAdapters;
-    int allAdaptersNum;
-    int position;
-
-    AdapterIterator(const std::vector<int>& _adapters, bool _allAdapters, int _allAdaptersNum) :
-        adapters(_adapters), allAdapters(_allAdapters), allAdaptersNum(_allAdaptersNum), position(0)
-    {
-
-    }
-
-    AdapterIterator& operator++()
-    {
-        position++;
-        return *this;
-    }
-
-    operator bool() const
-    {
-        return (!allAdapters && position < int(adapters.size())) || (allAdapters && position < allAdaptersNum);
-    }
-    bool operator!() const
-    {
-        return !((!allAdapters && position < int(adapters.size())) || (allAdapters && position < allAdaptersNum));
-    }
-    int operator*() const
-    {
-        return allAdapters ? position : adapters[position];
-    }
-};
-
-static void setOVCParameters(ADLMainControl& mainControl, int adaptersNum, const std::vector<int>& activeAdapters, const std::vector<OVCParameter>& ovcParams)
-{
-    std::cout << "WARNING: Setting AMD Overdrive parameters!" << std::endl;
-    std::cout <<
-        "\nIMPORTANT NOTICE: Before any setting of AMD Overdrive parameters,\nplease stop all GPU computations and GPU renderings.\n"
-        "Please use this utility carefully, as it can damage your hardware.\n" << std::endl;
-
-    const int realAdaptersNum = activeAdapters.size();
-
-    std::vector<ADLODParameters> odParams(realAdaptersNum);
-    std::vector<std::vector<ADLODPerformanceLevel> > perfLevels(realAdaptersNum);
-    std::vector<std::vector<ADLODPerformanceLevel> > defaultPerfLevels(realAdaptersNum);
-    std::vector<bool> changedDevices(realAdaptersNum);
-    std::fill(changedDevices.begin(), changedDevices.end(), false);
-
-    bool failed = false;
-
-    for (OVCParameter param: ovcParams)
-    {
-        if (!param.allAdapters)
-        {
-            bool listFailed = false;
-
-            for (int adapterIndex: param.adapters)
-            {
-                if (!listFailed && (adapterIndex>=realAdaptersNum || adapterIndex <0 ))
-                {
-                    std::cerr << "Some adapter indices are out of range in '" << param.argText << "'!" << std::endl;
-                    listFailed = failed = true;
-                }
-            }
-        }
-    }
-
-    // check fanspeed
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type == OVCParamType::FAN_SPEED)
-        {
-            if(param.partId != 0)
-            {
-                std::cerr << "Thermal Control Index is not 0 in '" << param.argText << "'!" << std::endl;
-                failed = true;
-            }
-            if(!param.useDefault && (param.value < 0.0 || param.value > 100.0))
-            {
-                std::cerr << "FanSpeed value out of range in '" << param.argText << "'!" << std::endl;
-                failed = true;
-            }
-        }
-    }
-
-    for (int ai = 0; ai < realAdaptersNum; ai++)
-    {
-        int i = activeAdapters[ai];
-
-        mainControl.getODParameters(i, odParams[ai]);
-
-        perfLevels[ai].resize(odParams[ai].iNumberOfPerformanceLevels);
-        defaultPerfLevels[ai].resize(odParams[ai].iNumberOfPerformanceLevels);
-
-        mainControl.getODPerformanceLevels(i, 0, odParams[ai].iNumberOfPerformanceLevels, perfLevels[ai].data());
-        mainControl.getODPerformanceLevels(i, 1, odParams[ai].iNumberOfPerformanceLevels, defaultPerfLevels[ai].data());
-    }
-
-    // check other params
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type!=OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, realAdaptersNum); ait; ++ait)
-            {
-                int i = *ait;
-
-                if (i>=realAdaptersNum)
-                {
-                    continue;
-                }
-
-                int partId = (param.partId!=LAST_PERFLEVEL) ? param.partId : odParams[i].iNumberOfPerformanceLevels - 1;
-
-                if (partId >= odParams[i].iNumberOfPerformanceLevels || partId < 0)
-                {
-                    std::cerr << "Performance level out of range in '" << param.argText << "'!" << std::endl;
-                    failed = true;
-                    continue;
-                }
-                switch(param.type)
-                {
-                    case OVCParamType::CORE_CLOCK:
-
-                        if (!param.useDefault && (param.value < odParams[i].sEngineClock.iMin/100.0 || param.value > odParams[i].sEngineClock.iMax/100.0))
-                        {
-                            std::cerr << "Core clock out of range in '" << param.argText << "'!" << std::endl;
-                            failed = true;
-                        }
-                        break;
-
-                    case OVCParamType::MEMORY_CLOCK:
-
-                        if (!param.useDefault && (param.value < odParams[i].sMemoryClock.iMin/100.0 || param.value > odParams[i].sMemoryClock.iMax/100.0))
-                        {
-                            std::cerr << "Memory clock out of range in '" << param.argText << "'!" << std::endl;
-                            failed = true;
-                        }
-                        break;
-
-                    case OVCParamType::VDDC_VOLTAGE:
-
-                        if (!param.useDefault && (param.value < odParams[i].sVddc.iMin/1000.0 || param.value > odParams[i].sVddc.iMax/1000.0))
-                        {
-                            std::cerr << "Voltage out of range in '" << param.argText << "'!" << std::endl;
-                            failed = true;
-                        }
-                        break;
-
-                    default:
-                        break;
-                }
-            }
-        }
-    }
-
-    if (failed)
-    {
-        std::cerr << "NO ANY settings applied. Error in parameters!" << std::endl;
-        throw Error("Wrong parameters!");
-    }
-
-    // print what has been changed
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type==OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, realAdaptersNum); ait; ++ait)
-            {
-                std::cout << "Setting fanspeed to ";
-
-                if (param.useDefault)
-                {
-                    std::cout << "default";
-                }
-                else
-                {
-                    std::cout << param.value << "%";
-                }
-
-                std::cout << " for adapter " << *ait << " at thermal controller " << param.partId << std::endl;
-            }
-        }
-    }
-
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type!=OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, realAdaptersNum); ait; ++ait)
-            {
-                int i = *ait;
-                int partId = (param.partId!=LAST_PERFLEVEL) ? param.partId: odParams[i].iNumberOfPerformanceLevels -1;
-
-                switch(param.type)
-                {
-                    case OVCParamType::CORE_CLOCK:
-
-                        std::cout << "Setting core clock to ";
-
-                        if (param.useDefault)
-                        {
-                            std::cout << "default";
-                        }
-                        else
-                        {
-                            std::cout << param.value << " MHz";
-                        }
-
-                        std::cout << " for adapter " << i << " at performance level " << partId << std::endl;
-                        break;
-
-                    case OVCParamType::MEMORY_CLOCK:
-
-                        std::cout << "Setting memory clock to ";
-
-                        if (param.useDefault)
-                        {
-                            std::cout << "default";
-                        }
-                        else
-                        {
-                            std::cout << param.value << " MHz";
-                        }
-
-                        std::cout << " for adapter " << i << " at performance level " << partId << std::endl;
-                        break;
-
-                    case OVCParamType::CORE_OD:
-
-                        std::cout << "Core OD available only for AMDGPU-(PRO) drivers." << std::endl;
-                        break;
-
-                    case OVCParamType::MEMORY_OD:
-
-                        std::cout << "Memory OD available only for AMDGPU-(PRO) drivers." << std::endl;
-                        break;
-
-                    case OVCParamType::VDDC_VOLTAGE:
-
-                        std::cout << "Setting Vddc voltage to ";
-
-                        if (param.useDefault)
-                        {
-                            std::cout << "default";
-                        }
-                        else
-                        {
-                            std::cout << param.value << " V";
-                        }
-
-                        std::cout << " for adapter " << i << " at performance level " << partId << std::endl;
-                        break;
-
-                    default:
-
-                        break;
-                }
-            }
-        }
-    }
-
-    std::vector<FanSpeedSetup> fanSpeedSetups(realAdaptersNum);
-    std::fill(fanSpeedSetups.begin(), fanSpeedSetups.end(), FanSpeedSetup{ 0.0, false, false });
-
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type==OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, realAdaptersNum); ait; ++ait)
-            {
-                fanSpeedSetups[*ait].value = param.value;
-                fanSpeedSetups[*ait].useDefault = param.useDefault;
-                fanSpeedSetups[*ait].isSet = true;
-            }
-        }
-    }
-
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type!=OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, realAdaptersNum); ait; ++ait)
-            {
-                int i = *ait;
-                int partId = (param.partId != LAST_PERFLEVEL) ? param.partId : odParams[i].iNumberOfPerformanceLevels - 1;
-                ADLODPerformanceLevel& perfLevel = perfLevels[i][partId];
-                const ADLODPerformanceLevel& defaultPerfLevel = defaultPerfLevels[i][partId];
-
-                switch(param.type)
-                {
-                    case OVCParamType::CORE_CLOCK:
-
-                        if (param.useDefault)
-                        {
-                            perfLevel.iEngineClock = defaultPerfLevel.iEngineClock;
-                        }
-                        else
-                        {
-                            perfLevel.iEngineClock = int(round(param.value * 100.0));
-                        }
-
-                        break;
-
-                    case OVCParamType::MEMORY_CLOCK:
-
-                        if (param.useDefault)
-                        {
-                            perfLevel.iMemoryClock = defaultPerfLevel.iMemoryClock;
-                        }
-                        else
-                        {
-                            perfLevel.iMemoryClock = int(round(param.value * 100.0));
-                        }
-
-                        break;
-
-                    case OVCParamType::VDDC_VOLTAGE:
-
-                        if (param.useDefault)
-                        {
-                            perfLevel.iVddc = defaultPerfLevel.iVddc;
-                        }
-                        else if (perfLevel.iVddc==0)
-                        {
-                            std::cout << "Voltage for adapter " << i << " is not set!" << std::endl;
-                        }
-                        else
-                        {
-                            perfLevel.iVddc = int(round(param.value * 1000.0));
-                        }
-                        break;
-
-                    default:
-
-                        break;
-                }
-
-                changedDevices[i] = true;
-            }
-        }
-    }
-
-    /// set fan speeds
-    for (int i = 0; i < realAdaptersNum; i++)
-    {
-        if (fanSpeedSetups[i].isSet)
-        {
-            if (!fanSpeedSetups[i].useDefault)
-            {
-                mainControl.setFanSpeed(activeAdapters[i], 0 /* must be zero */, int(round(fanSpeedSetups[i].value)));
-            }
-            else
-            {
-                mainControl.setFanSpeedToDefault(activeAdapters[i], 0);
-            }
-        }
-    }
-
-    // set od perflevels
-    for (int i = 0; i < realAdaptersNum; i++)
-    {
-        if (changedDevices[i])
-        {
-            mainControl.setODPerformanceLevels(activeAdapters[i], odParams[i].iNumberOfPerformanceLevels, perfLevels[i].data());
-        }
-    }
-}
-
-/* AMDGPU code */
-
-struct PerfClocks
-{
-    unsigned int coreClock;
-    unsigned int memoryClock;
-};
-
-static void setOVCParameters(AMDGPUAdapterHandle& handle, const std::vector<OVCParameter>& ovcParams, const std::vector<PerfClocks>& perfClocks)
-{
-    std::cout << "WARNING: setting AMD Overdrive parameters!" << std::endl;
-    std::cout << "\nIMPORTANT NOTICE: Before any setting of AMD Overdrive parameters,\nplease STOP ANY GPU computations and GPU renderings.\n"
-        "Please use this utility CAREFULLY, because it can DAMAGE your hardware!\n" << std::endl;
-
-    bool failed = false;
-    int adaptersNum = handle.getAdaptersNum();
-
-    for (OVCParameter param: ovcParams)
-    {
-        if (!param.allAdapters)
-        {
-            bool listFailed = false;
-            for (int adapterIndex: param.adapters)
-            {
-                if (!listFailed && (adapterIndex>=adaptersNum || adapterIndex < 0))
-                {
-                    std::cerr << "Some adapter indices out of range in '" << param.argText << "'!" << std::endl;
-                    listFailed = failed = true;
-                }
-            }
-        }
-    }
-
-    // check fanspeed
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type == OVCParamType::FAN_SPEED)
-        {
-            if(param.partId != 0)
-            {
-                std::cerr << "Thermal Control Index is not 0 in '" << param.argText << "'!" << std::endl;
-                failed = true;
-            }
-            if(!param.useDefault && (param.value < 0.0 || param.value>100.0))
-            {
-                std::cerr << "FanSpeed value out of range in '" << param.argText << "'!" << std::endl;
-                failed = true;
-            }
-        }
-    }
-
-    // check other params
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type!=OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, adaptersNum); ait; ++ait)
-            {
-                int i = *ait;
-                if (i>=adaptersNum)
-                {
-                    continue;
-                }
-
-                int partId = (param.partId != LAST_PERFLEVEL) ? param.partId : 0;
-
-                if (partId != 0)
-                {
-                    std::cerr << "Performance level out of range in '" << param.argText << "'!" << std::endl;
-                    failed = true;
-                    continue;
-                }
-
-                const PerfClocks& perfClks = perfClocks[i];
-
-                switch(param.type)
-                {
-                    case OVCParamType::CORE_CLOCK:
-
-                        if (!param.useDefault && (param.value < perfClks.coreClock || param.value > perfClks.coreClock * 1.20))
-                        {
-                            std::cerr << "Core clock out of range in '" << param.argText << "'!" << std::endl;
-                            failed = true;
-                        }
-                        break;
-
-                    case OVCParamType::MEMORY_CLOCK:
-
-                        if (!param.useDefault && (param.value < perfClks.memoryClock || param.value > perfClks.memoryClock * 1.20))
-                        {
-                            std::cerr << "Memory clock out of range in '" << param.argText << "'!" << std::endl;
-                            failed = true;
-                        }
-                        break;
-
-                    case OVCParamType::CORE_OD:
-
-                        if (!param.useDefault && (param.value < 0.0 || param.value > 20.0))
-                        {
-                            std::cerr << "Core Overdrive out of range in '" << param.argText << "'!" << std::endl;
-                            failed = true;
-                        }
-                        break;
-
-                    case OVCParamType::MEMORY_OD:
-
-                        if (!param.useDefault && (param.value < 0.0 || param.value > 20.0))
-                        {
-                            std::cerr << "Memory Overdrive out of range in '" << param.argText << "'!" << std::endl;
-                            failed = true;
-                        }
-                        break;
-
-                    default:
-
-                        break;
-                }
-            }
-        }
-    }
-
-    if (failed)
-    {
-        std::cerr << "Error in parameters. No settings have been applied." << std::endl;
-        throw Error("Invalid parameters.");
-    }
-
-    // print what has been changed
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type==OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, adaptersNum); ait; ++ait)
-            {
-                std::cout << "Setting fan speed to ";
-
-                if (param.useDefault)
-                {
-                    std::cout << "default";
-                }
-                else
-                {
-                    std::cout << param.value << "%";
-                }
-
-                std::cout << " for adapter " << *ait << " at thermal controller " << param.partId << std::endl;
-            }
-        }
-    }
-
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type!=OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, adaptersNum); ait; ++ait)
-            {
-                int i = *ait;
-                int partId = (param.partId!=LAST_PERFLEVEL)?param.partId:0;
-
-                switch(param.type)
-                {
-                    case OVCParamType::CORE_CLOCK:
-
-                        std::cout << "Setting core clock to ";
-                        if (param.useDefault)
-                        {
-                            std::cout << "default";
-                        }
-                        else
-                        {
-                            std::cout << param.value << " MHz";
-                        }
-                        std::cout << " for adapter " << i << " at performance level " << partId << std::endl;
-                        break;
-
-                    case OVCParamType::MEMORY_CLOCK:
-
-                        std::cout << "Setting memory clock to ";
-                        if (param.useDefault)
-                        {
-                            std::cout << "default";
-                        }
-                        else
-                        {
-                            std::cout << param.value << " MHz";
-                        }
-                        std::cout << " for adapter " << i << " at performance level " << partId << std::endl;
-                        break;
-
-                    case OVCParamType::CORE_OD:
-
-                        std::cout << "Setting core overdrive to ";
-
-                        if (param.useDefault)
-                        {
-                            std::cout << "default";
-                        }
-                        else
-                        {
-                            std::cout << param.value;
-                        }
-
-                        std::cout << " for adapter " << i << " at performance level " << partId << std::endl;
-                        break;
-
-                    case OVCParamType::MEMORY_OD:
-
-                        std::cout << "Setting memory overdrive to ";
-
-                        if (param.useDefault)
-                        {
-                            std::cout << "default";
-                        }
-                        else
-                        {
-                            std::cout << param.value;
-                        }
-
-                        std::cout << " for adapter " << i << " at performance level " << partId << std::endl;
-                        break;
-
-                    case OVCParamType::VDDC_VOLTAGE:
-
-                        std::cout << "VDDC voltage available only for AMD Catalyst/Crimson drivers." << std::endl;
-                        break;
-
-                    default:
-
-                        break;
-                }
-            }
-        }
-    }
-
-    std::vector<FanSpeedSetup> fanSpeedSetups(adaptersNum);
-    std::fill(fanSpeedSetups.begin(), fanSpeedSetups.end(), FanSpeedSetup{ 0.0, false, false });
-
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type==OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, adaptersNum); ait; ++ait)
-            {
-                fanSpeedSetups[*ait].value = param.value;
-                fanSpeedSetups[*ait].useDefault = param.useDefault;
-                fanSpeedSetups[*ait].isSet = true;
-            }
-        }
-    }
-
-    for (OVCParameter param: ovcParams)
-    {
-        if (param.type!=OVCParamType::FAN_SPEED)
-        {
-            for (AdapterIterator ait(param.adapters, param.allAdapters, adaptersNum); ait; ++ait)
-            {
-                int i = *ait;
-
-                const PerfClocks& perfClks = perfClocks[i];
-
-                switch(param.type)
-                {
-                    case OVCParamType::CORE_CLOCK:
-
-                        if (param.useDefault)
-                        {
-                            handle.setOverdriveCoreParam(i, 0);
-                        }
-                        else
-                        {
-                            handle.setOverdriveCoreParam(i, int(round((double(param.value - perfClks.coreClock) / perfClks.coreClock) * 100.0)));
-                        }
-                        break;
-
-                    case OVCParamType::MEMORY_CLOCK:
-
-                        if (param.useDefault)
-                        {
-                            handle.setOverdriveMemoryParam(i, 0);
-                        }
-                        else
-                        {
-                            handle.setOverdriveMemoryParam(i, int(round((double(param.value - perfClks.memoryClock) / perfClks.memoryClock) * 100.0)));
-                        }
-                        break;
-
-                    case OVCParamType::CORE_OD:
-
-                        if (param.useDefault)
-                        {
-                            handle.setOverdriveCoreParam(i, 0);
-                        }
-                        else
-                        {
-                            handle.setOverdriveCoreParam(i, int(round(param.value)));
-                        }
-                        break;
-
-                    case OVCParamType::MEMORY_OD:
-
-                        if (param.useDefault)
-                        {
-                            handle.setOverdriveMemoryParam(i, 0);
-                        }
-                        else
-                        {
-                            handle.setOverdriveMemoryParam(i, int(round(param.value)));
-                        }
-                        break;
-
-                    default:
-
-                        break;
-                }
-            }
-        }
-    }
-
-    /// set fan speeds
-    for (int i = 0; i < adaptersNum; i++)
-    {
-        if (fanSpeedSetups[i].isSet)
-        {
-            if (!fanSpeedSetups[i].useDefault)
-            {
-                handle.setFanSpeed(i, int(round(fanSpeedSetups[i].value)));
-            }
-            else
-            {
-                handle.setFanSpeedToDefault(i);
-            }
-        }
-    }
-}
-
-/* AMDGPU code */
 
 static const char* helpAndUsageString =
     "amdcovc " AMDCOVC_VERSION " by Mateusz Szpakowski (matszpk@interia.pl)\n"
@@ -1442,6 +127,12 @@ static const char* helpAndUsageString =
     "\n"
     "If the X11 server is not running, then this program requires root privileges.\n";
 
+static const char* versionString =
+    "amdcovc " AMDCOVC_VERSION
+    " by Mateusz Szpakowski (matszpk@interia.pl)\n"
+    "Program is distributed under terms of the GPLv2.\n"
+    "Program available at https://github.com/matszpk/amdcovc.\n"
+
 int main(int argc, const char** argv)
 try
 {
@@ -1466,14 +157,14 @@ try
         }
         else if (::strncmp(argv[i], "--adapters=", 11) == 0)
         {
-            parseAdaptersList(argv[i] + 11, choosenAdapters, chooseAllAdapters);
+            Adapters::ParseAdaptersList(argv[i] + 11, choosenAdapters, chooseAllAdapters);
             useAdaptersList = true;
         }
         else if (::strcmp(argv[i], "--adapters") == 0)
         {
             if ( i + 1 < argc)
             {
-                parseAdaptersList(argv[++i], choosenAdapters, chooseAllAdapters);
+                Adapters::ParseAdaptersList(argv[++i], choosenAdapters, chooseAllAdapters);
                 useAdaptersList = true;
             }
             else
@@ -1485,11 +176,11 @@ try
         {
             if (argv[i][2]!=0)
             {
-                parseAdaptersList(argv[i]+2, choosenAdapters, chooseAllAdapters);
+                Adapters::ParseAdaptersList(argv[i] + 2, choosenAdapters, chooseAllAdapters);
             }
             else if (i+1 < argc)
             {
-                parseAdaptersList(argv[++i], choosenAdapters, chooseAllAdapters);
+                Adapters::ParseAdaptersList(argv[++i], choosenAdapters, chooseAllAdapters);
             }
             else
             {
@@ -1500,18 +191,15 @@ try
         }
         else if (::strcmp(argv[i], "--version") == 0)
         {
-            std::cout << "amdcovc " AMDCOVC_VERSION
-            " by Mateusz Szpakowski (matszpk@interia.pl)\n"
-            "Program is distributed under terms of the GPLv2.\n"
-            "Program available at https://github.com/matszpk/amdcovc.\n" << std::endl;
-
+            std::cout << versionString;
+            std::cout.flush();
             return 0;
         }
         else
         {
             OVCParameter param;
 
-            if (parseOVCParameter(argv[i], param))
+            if (Parameters::ParseOVCParameter(argv[i], param))
             {
                 ovcParameters.push_back(param);
             }
@@ -1526,7 +214,6 @@ try
     {
         std::cout << helpAndUsageString;
         std::cout.flush();
-
         return 0;
     }
 
@@ -1545,7 +232,7 @@ try
 
         // list for converting user indices to input indices to ADL interface
         std::vector<int> activeAdapters;
-        getActiveAdaptersIndices(mainControl, adaptersNum, activeAdapters);
+        Adapters::GetActiveAdaptersIndices(mainControl, adaptersNum, activeAdapters);
 
         if (useAdaptersList)
         {
@@ -1561,17 +248,17 @@ try
 
         if (!ovcParameters.empty())
         {
-            setOVCParameters(mainControl, adaptersNum, activeAdapters, ovcParameters);
+            Parameters::SetOVCParameters(mainControl, adaptersNum, activeAdapters, ovcParameters);
         }
         else
         {
             if (printVerbose)
             {
-                printAdaptersInfoVerbose(mainControl, adaptersNum, activeAdapters, choosenAdapters, useAdaptersList && !chooseAllAdapters);
+                Adapters::PrintAdaptersInfoVerbose(mainControl, adaptersNum, activeAdapters, choosenAdapters, useAdaptersList && !chooseAllAdapters);
             }
             else
             {
-                printAdaptersInfo(mainControl, adaptersNum, activeAdapters, choosenAdapters, useAdaptersList && !chooseAllAdapters);
+                Adapters::PrintAdaptersInfo(mainControl, adaptersNum, activeAdapters, choosenAdapters, useAdaptersList && !chooseAllAdapters);
             }
         }
     }
@@ -1591,7 +278,7 @@ try
                 perfClocks.push_back(PerfClocks{ coreClock, memoryClock });
             }
 
-            setOVCParameters(handle, ovcParameters, perfClocks);
+            Parameters::SetOVCParameters(handle, ovcParameters, perfClocks);
         }
         else
         {
@@ -1609,11 +296,11 @@ try
 
             if (printVerbose)
             {
-                printAdaptersInfoVerbose(handle, choosenAdapters, useAdaptersList && !chooseAllAdapters);
+                Adapters::PrintAdaptersInfoVerbose(handle, choosenAdapters, useAdaptersList && !chooseAllAdapters);
             }
             else
             {
-                printAdaptersInfo(handle, choosenAdapters, useAdaptersList && !chooseAllAdapters);
+                Adapters::PrintAdaptersInfo(handle, choosenAdapters, useAdaptersList && !chooseAllAdapters);
             }
         }
     }
